requirements.txt
absl-py==2.2.1
attrs==25.3.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.21
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
protobuf==4.25.6
pycparser==2.22
pyparsing==3.2.3
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
Werkzeug==3.1.3
